refactor : DeleteHistories, DeleteHistory - DeleteHistories : 컨벤션에 따라 메서드 네이밍 변경 (of() -> get()) - DeleteHistory : 각 ContentType에 따른 DeleteHistory 처리를 DeleteHistories에서 DeleteHistory 책임으로 위임하여 create()로 처리
